Enhance waitUntilTextExists() to find substrings in case the selector matches multiple elements
Until now substring matching was only working if the selector matched a single element.
If the selector matches multiple elements getText() returns an array instead of a string.
Using indexOf() on an array compares only the full array entry not a substring.
Therefore this case has to be handled especially.